Made FileUtil.delete() not throw an error if the file/folder didn't exist in the first place

+++ common/src/com/redhat/ceylon/common/FileUtil.java
@@ -29,13 +29,15 @@ public class FileUtil {
      * @throws RuntimeException if the file or directory could not be deleted
      */
     public static void delete(File f){
-        if (f.isDirectory()) {
-            for (File c : f.listFiles()) {
-                delete(c);
+        if (f.exists()) {
+            if (f.isDirectory()) {
+                for (File c : f.listFiles()) {
+                    delete(c);
+                }
             }
-        }
-        if (!f.delete()) {
-            throw new RuntimeException("Failed to delete file: " + f.getPath());
+            if (!f.delete()) {
+                throw new RuntimeException("Failed to delete file: " + f.getPath());
+            }
         }
     }
